+++ include/utilidades.h
@@ -12,8 +12,10 @@
 
 void binarioNaTela(char *nomeArquivoBinario);
 void scan_quote_string(char *str);
-void atualizaAVL(ARV* arv, int id, long long int bOffset);
+
+void atualizaAVL(ARV* arv, int id, long int bOffset);
 void printCrescIndice(NO* raiz, FILE* indice_bin_file);
+
 char *novo_strtok(char *str, const char *delim, char **saveptr);
 char *trim(char *str);
